+++ dataprovider/basketballreference/nba/basic_box_score_stats.go
@@ -82,7 +82,7 @@ func getBasicBoxScoreStats(nbaMatchup interface{}) []interface{} {
 	start := time.Now().UTC()
 	var basicNBABoxScoreStats []interface{}
 	fmt.Println("Scraping Basic Box Score: " + url)
-	dr := request.NewDocumentRetriever(2 * time.Minute)
+	dr := request.NewDocumentRetriever(request.WithTimeout(2 * time.Minute))
 	doc, err := dr.RetrieveDocument(url, networkHeaders, waitReadyBoxScoreContentSelector)
 	if err != nil {
 		log.Fatalln(err)